Refactor injection and strategy pattern
